refactor: remove useless files, apply bootstrap and bootswatch themes
