+++ src/components/App/index.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import logo from '../../assets/img/logo.png';
+import './style.scss';
+
+function App() {
+  return (
+    <div className="header d-flex">
+        <img src={logo} alt="logo" />
+    </div>
+  );
+}
+
+export default App;
