+++ src/components/App/style.scss
